test: align draft repro with browser guidance
Amp-Thread-ID: https://ampcode.com/threads/T-01a087ba-8814-77b1-9137-17e16a298de0
